Add files via upload
Eon Spar is a JRPG-style roguelike made in PICO-8, which is a dialect of Lua, a language similar to C. To run the program, go to https://www.pico-8-edu.com/, hit the play button, and drag and drop Eon Spar.p8 from your file manager onto the browser. Then, hit Esc to look at the code and sprites, and hit Ctrl+R to run the code.